# executor: try to optimize index hash join query by reduce cop task count