<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #54840

Problem Summary:  try to optimize index hash join query by reduce cop task count.

### What changed and how does it work?

Increate index hash join outer executor initial batch size.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

1. prepare data:

```sql
drop table if exists t0,t1;
create table t1 (id int key auto_increment, b int, index idx1 (b));
insert into t1 () values (), (), (), (), (), (), (), ();
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1 limit 34464;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1;
insert into t1 (b) select b from t1 limit 200000;
update t1 set b = rand() * 100000;
create table t0 (id int key, b int, index idx1 (b));
insert into t0 select * from t1 order by id limit 100000;
update t0 set b = rand() * 100000;
split table t1 index idx1 between (0) and (100000) regions 10;
analyze table t0,t1;
set @@tidb_store_batch_size=0;
```

**This PR**

- Pay attention to the execution information of `IndexReader_11` executor's `cop_task` number, it is less than before by default.
- When increasing `tidb_max_chunk_size`, the `IndexReader_11` executor's `cop_task` number decreases. 
- When increasing `tidb_max_chunk_size`, the `IndexReader_11` executor's `actRows` decreases, this reduces the stress on the tikv.

```sql
-- IndexReader_11's cop_task num is 309
> explain analyze select /*+ INL_HASH_JOIN(t0,t1) */ t0.* from t0 join t1 where t0.id<=50000 and t0.b=t1.b;
+-------------------------+-----------+---------+-----------+-------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+
| id                      | estRows   | actRows | task      | access object           | execution info                                                                                                                                                                                                                                                                                                                                                                                                                                                                   | operator info                                                                                                   | memory   | disk |
+-------------------------+-----------+---------+-----------+-------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+
| IndexHashJoin_14        | 503451.52 | 503086  | root      |                         | time:188.7ms, loops:495, RU:899.657452, inner:{total:456.9ms, concurrency:5, task:5, construct:18.6ms, fetch:359.2ms, build:3.54ms, join:79.2ms}                                                                                                                                                                                                                                                                                                                                 | inner join, inner:IndexReader_11, outer key:test.t0.b, inner key:test.t1.b, equal cond:eq(test.t0.b, test.t1.b) | 11.1 MB  | N/A  |
| ├─TableReader_33(Build) | 50022.94  | 50000   | root      |                         | time:41.3ms, loops:52, cop_task: {num: 9, max: 14.1ms, min: 462.2µs, avg: 4.6ms, p95: 14.1ms, max_proc_keys: 17376, p95_proc_keys: 17376, tot_proc: 23.4ms, tot_wait: 2.63ms, copr_cache: disabled, build_task_duration: 4.75µs, max_distsql_concurrency: 1}, rpc_info:{Cop:{num_rpc:9, total_time:41.4ms}}                                                                                                                                                                      | data:Selection_32                                                                                               | 453.9 KB | N/A  |
| │ └─Selection_32        | 50022.94  | 50000   | cop[tikv] |                         | tikv_task:{proc max:13ms, min:0s, avg: 3.89ms, p80:9ms, p95:13ms, iters:84, tasks:9}, scan_detail: {total_process_keys: 50000, total_process_keys_size: 1967413, total_keys: 100007, get_snapshot_time: 139.3µs, rocksdb: {key_skipped_count: 99998, block: {cache_hit_count: 161}}}, time_detail: {total_process_time: 23.4ms, total_suspend_time: 12.5ms, total_wait_time: 2.63ms, total_kv_read_wall_time: 33ms, tikv_wall_time: 39.4ms}                                      | not(isnull(test.t0.b))                                                                                          | N/A      | N/A  |
| │   └─TableRangeScan_31 | 50022.94  | 50000   | cop[tikv] | table:t0                | tikv_task:{proc max:13ms, min:0s, avg: 3.67ms, p80:8ms, p95:13ms, iters:84, tasks:9}                                                                                                                                                                                                                                                                                                                                                                                             | range:[-inf,50000], keep order:false                                                                            | N/A      | N/A  |
| └─IndexReader_11(Probe) | 503451.52 | 469361  | root      |                         | time:339.7ms, loops:487, cop_task: {num: 309, max: 30.9ms, min: 892.3µs, avg: 10.3ms, p95: 26.9ms, max_proc_keys: 5088, p95_proc_keys: 5088, tot_proc: 1.14s, tot_wait: 386.2ms, copr_cache: disabled, build_task_duration: 1.43ms, max_distsql_concurrency: 11}, rpc_info:{Cop:{num_rpc:309, total_time:3.2s}}                                                                                                                                                                  | index:Selection_10                                                                                              | 16.8 KB  | N/A  |
|   └─Selection_10        | 503451.52 | 469361  | cop[tikv] |                         | tikv_task:{proc max:29ms, min:0s, avg: 8.26ms, p80:17ms, p95:25ms, iters:1624, tasks:309}, scan_detail: {total_process_keys: 469361, total_process_keys_size: 21590606, total_keys: 516301, get_snapshot_time: 6.01ms, rocksdb: {delete_skipped_count: 3134, key_skipped_count: 472495, block: {cache_hit_count: 207274}}}, time_detail: {total_process_time: 1.14s, total_suspend_time: 1.42s, total_wait_time: 386.2ms, total_kv_read_wall_time: 2.54s, tikv_wall_time: 3.05s} | not(isnull(test.t1.b))                                                                                          | N/A      | N/A  |
|     └─IndexRangeScan_9  | 503451.52 | 469361  | cop[tikv] | table:t1, index:idx1(b) | tikv_task:{proc max:29ms, min:0s, avg: 8.21ms, p80:17ms, p95:25ms, iters:1624, tasks:309}                                                                                                                                                                                                                                                                                                                                                                                        | range: decided by [eq(test.t1.b, test.t0.b)], keep order:false                                                  | N/A      | N/A  |
+-------------------------+-----------+---------+-----------+-------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+

> set @@tidb_max_chunk_size=4096;
-- IndexReader_11's cop_task num is 222.
> explain analyze select /*+ INL_HASH_JOIN(t0,t1) */ t0.* from t0 join t1 where t0.id<=50000 and t0.b=t1.b;
+-------------------------+-----------+---------+-----------+-------------------------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+
| id                      | estRows   | actRows | task      | access object           | execution info                                                                                                                                                                                                                                                                                                                                                                                                                                                                  | operator info                                                                                                   | memory   | disk |
+-------------------------+-----------+---------+-----------+-------------------------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+
| IndexHashJoin_14        | 503451.52 | 503086  | root      |                         | time:219.5ms, loops:125, RU:831.027614, inner:{total:396.4ms, concurrency:5, task:3, construct:29.2ms, fetch:289.9ms, build:3.87ms, join:77.3ms}                                                                                                                                                                                                                                                                                                                                | inner join, inner:IndexReader_11, outer key:test.t0.b, inner key:test.t1.b, equal cond:eq(test.t0.b, test.t1.b) | 12.4 MB  | N/A  |
| ├─TableReader_33(Build) | 50022.94  | 50000   | root      |                         | time:33.3ms, loops:14, cop_task: {num: 9, max: 11.7ms, min: 540.2µs, avg: 3.7ms, p95: 11.7ms, max_proc_keys: 17376, p95_proc_keys: 17376, tot_proc: 23.7ms, tot_wait: 1.5ms, copr_cache: disabled, build_task_duration: 11.5µs, max_distsql_concurrency: 1}, rpc_info:{Cop:{num_rpc:9, total_time:33.2ms}}                                                                                                                                                                      | data:Selection_32                                                                                               | 453.9 KB | N/A  |
| │ └─Selection_32        | 50022.94  | 50000   | cop[tikv] |                         | tikv_task:{proc max:10ms, min:0s, avg: 3ms, p80:7ms, p95:10ms, iters:84, tasks:9}, scan_detail: {total_process_keys: 50000, total_process_keys_size: 1967413, total_keys: 100007, get_snapshot_time: 230.9µs, rocksdb: {key_skipped_count: 99998, block: {cache_hit_count: 161}}}, time_detail: {total_process_time: 23.7ms, total_suspend_time: 5.58ms, total_wait_time: 1.5ms, total_kv_read_wall_time: 27ms, tikv_wall_time: 31.4ms}                                         | not(isnull(test.t0.b))                                                                                          | N/A      | N/A  |
| │   └─TableRangeScan_31 | 50022.94  | 50000   | cop[tikv] | table:t0                | tikv_task:{proc max:10ms, min:0s, avg: 3ms, p80:7ms, p95:10ms, iters:84, tasks:9}                                                                                                                                                                                                                                                                                                                                                                                               | range:[-inf,50000], keep order:false                                                                            | N/A      | N/A  |
| └─IndexReader_11(Probe) | 503451.52 | 456277  | root      |                         | time:264.7ms, loops:116, cop_task: {num: 222, max: 34.5ms, min: 1.32ms, avg: 11.4ms, p95: 27.6ms, max_proc_keys: 9184, p95_proc_keys: 5088, tot_proc: 1.09s, tot_wait: 212.8ms, copr_cache: disabled, build_task_duration: 3.96ms, max_distsql_concurrency: 11}, rpc_info:{Cop:{num_rpc:222, total_time:2.53s}}                                                                                                                                                                 | index:Selection_10                                                                                              | 49.5 KB  | N/A  |
|   └─Selection_10        | 503451.52 | 456277  | cop[tikv] |                         | tikv_task:{proc max:32ms, min:0s, avg: 9.05ms, p80:18ms, p95:26ms, iters:1298, tasks:222}, scan_detail: {total_process_keys: 456277, total_process_keys_size: 20988742, total_keys: 501839, get_snapshot_time: 4.13ms, rocksdb: {delete_skipped_count: 3089, key_skipped_count: 459366, block: {cache_hit_count: 201212}}}, time_detail: {total_process_time: 1.09s, total_suspend_time: 928.5ms, total_wait_time: 212.8ms, total_kv_read_wall_time: 2s, tikv_wall_time: 2.35s} | not(isnull(test.t1.b))                                                                                          | N/A      | N/A  |
|     └─IndexRangeScan_9  | 503451.52 | 456277  | cop[tikv] | table:t1, index:idx1(b) | tikv_task:{proc max:32ms, min:0s, avg: 9.01ms, p80:18ms, p95:25ms, iters:1298, tasks:222}                                                                                                                                                                                                                                                                                                                                                                                       | range: decided by [eq(test.t1.b, test.t0.b)], keep order:false                                                  | N/A      | N/A  |
+-------------------------+-----------+---------+-----------+-------------------------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+


> set @@tidb_max_chunk_size=25000;
-- IndexReader_11's cop_task num is 168
> explain analyze select /*+ INL_HASH_JOIN(t0,t1) */ t0.* from t0 join t1 where t0.id<=50000 and t0.b=t1.b;
+-------------------------+-----------+---------+-----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------+----------+------+
| id                      | estRows   | actRows | task      | access object           | execution info                                                                                                                                                                                                                                                                                                                                                                                                                                                                     | operator info                                                                                                   | memory   | disk |
…