@crazycs520 2024-07-30
The tpch test results are as follows, and the PR will be slightly faster.

bench_type | tidb_version | QPM | total_avg_elapsed_time(s) | duration | tidb_commit_hash
-- | -- | -- | -- | -- | --
tpch | 8.2.0-cs | 1.26 | 544 | 2433 | fa8f14b15b8a439985c9f335ade84cab84b95d7a
tpch | 8.3.0-alpha-6-g2ee8c99 | 1.35 | 519 | 2406 | 2ee8c99cff49684b67b992a43646c1c3bc06b1e2